first pass at removing redundant details from dockerfile and beginning code reuse
