+++ entity-counter-recalculator/Dockerfile
@@ -1,18 +1,7 @@
-# Stage 1 - grab the image
-
-FROM python:3.11
-LABEL authors="JackLewis"
+# Stage 1 - build function and dependencies
+FROM python:3.11 AS build-image
 
 RUN apt-get update
-
-# Stage 2 - build function and dependencies
-FROM python AS build-image
-
-# Create function directory
-RUN mkdir -p /home/app/
-# Copy handler function
-COPY main.py /home/app/
-COPY /app /home/app/app
 
 # Install Lambda Runtime Interface Client for Python
 RUN pip install awslambdaric --target /home/app/
@@ -22,9 +11,15 @@ COPY minimalRequirements.txt ./
 RUN pip install --no-cache-dir --upgrade pip \
   && pip install --no-cache-dir -r minimalRequirements.txt --target /home/app/
 
-# Stage 3 - final runtime image
+# Create function directory
+RUN mkdir -p /home/app/
+# Copy handler function
+COPY main.py /home/app/
+COPY /app /home/app/app
+
+# Stage 2 - final runtime image
 # Grab a fresh copy of the Python image
-FROM python
+FROM python:3.11
 # Set working directory to function root directory
 WORKDIR /home/app
 # Copy in the built dependencies
