+++ entity-counter-recalculator/app/settings.py
@@ -4,14 +4,13 @@ import os
 def _get_boolean(env_name: str, fallback: str) -> bool:
     return os.environ.get(env_name, fallback).lower() in ("true", "t", "1")
 
-
 # AWS
 REGION = os.environ.get("AWS_REGION", "eu-west-1")
 ENABLE_CLOUDWATCH_INTEGRATION = _get_boolean("ENABLE_CLOUDWATCH_INTEGRATION", "True")
-CLOUDWATCH_CUSTOMER_DIFFERENCE_METRIC_NAME = os.environ.get("CLOUDWATCH_CUSTOMER_DIFFERENCE_METRIC_NAME")
-CLOUDWATCH_SPACE_DIFFERENCE_METRIC_NAME = os.environ.get("CLOUDWATCH_SPACE_DIFFERENCE_METRIC_NAME")
-CLOUDWATCH_CUSTOMER_DELETE_METRIC_NAME = os.environ.get("CLOUDWATCH_CUSTOMER_DELETE_METRIC_NAME")
-CLOUDWATCH_SPACE_DELETE_METRIC_NAME = os.environ.get("CLOUDWATCH_SPACE_DELETE_METRIC_NAME")
+CLOUDWATCH_CUSTOMER_DIFFERENCE_METRIC_NAME = os.environ.get("CLOUDWATCH_CUSTOMER_DIFFERENCE_METRIC_NAME", "CustomerTotalDelta")
+CLOUDWATCH_SPACE_DIFFERENCE_METRIC_NAME = os.environ.get("CLOUDWATCH_SPACE_DIFFERENCE_METRIC_NAME", "SpaceTotalDelta")
+CLOUDWATCH_CUSTOMER_DELETE_METRIC_NAME = os.environ.get("CLOUDWATCH_CUSTOMER_DELETE_METRIC_NAME", "SpaceDeleteMetrics")
+CLOUDWATCH_SPACE_DELETE_METRIC_NAME = os.environ.get("CLOUDWATCH_SPACE_DELETE_METRIC_NAME", "CustomerDeleteMetrics")
 APP_VERSION = os.environ.get('APP_VERSION', "1.0")
 AWS_CONNECTION_STRING_LOCATION = os.environ.get("AWS_CONNECTION_STRING_LOCATION")
